reorganized message handler

+++ events/message.js
@@ -14,8 +14,8 @@ module.exports = (client, message) => {
             if (err)
                 return client.log.error(err);
         });
-    } 
-    
+    }
+
     if (message.channel.name.startsWith('rs-')) {
         try {
             if (!fs.existsSync(`./tickets/${message.channel.name}.txt`)) {
@@ -58,33 +58,35 @@ module.exports = (client, message) => {
         });
     }
     // filter messages for commands
-    if (message.content.indexOf(client.config.prefix) !== 0) {
-        if (message.channel.name.startsWith('registration-')) {
-            return message.delete();
-        } else if (message.channel.name.startsWith('bot-')) {
-            return message.delete();
-        } else return;
-    };
+    if (msg.content.startsWith(client.config.prefix)) {
+        if (message.content.indexOf(client.config.prefix) !== 0) {
+            if (message.channel.name.startsWith('registration-')) {
+                return message.delete();
+            } else if (message.channel.name.startsWith('bot-')) {
+                return message.delete();
+            } else return;
+        };
 
-    // command prefix and formatting
-    // const args = message.content.slice(client.config.prefix.length).trim().split(/"/);
-    let input = message;
-    input = input.toString();
-    const regex = new RegExp('"[^"]+"|[\\S]+', 'g');
-    const args = [];
-    input.match(regex).forEach(element => {
-        if (!element) return;
-        return args.push(element.replace(/"/g, ''));
-    });
+        // command prefix and formatting
+        // const args = message.content.slice(client.config.prefix.length).trim().split(/"/);
+        let input = message;
+        input = input.toString();
+        const regex = new RegExp('"[^"]+"|[\\S]+', 'g');
+        const args = [];
+        input.match(regex).forEach(element => {
+            if (!element) return;
+            return args.push(element.replace(/"/g, ''));
+        });
 
-    let command = args.shift().toLowerCase();
-    command = command.slice(client.config.prefix.length);
-    client.log.console(`${message.author.tag} used the '${command}' command`);
-    client.log.console(args);
+        let command = args.shift().toLowerCase();
+        command = command.slice(client.config.prefix.length);
+        client.log.console(`${message.author.tag} used the '${command}' command`);
+        client.log.console(args);
 
-    // Grab the command data from the client.commands Enmap
-    const cmd = client.commands.get(command);
+        // Grab the command data from the client.commands Enmap
+        const cmd = client.commands.get(command);
 
-    // Run the command
-    cmd.run(client, message, args);
+        // Run the command
+        cmd.run(client, message, args);
+    }
 };
